Move .service classes to the .events.peopleevents package
