+++ android_mobile/src/test/java/com/alexstyl/specialdates/events/peopleevents/PeopleEventsProviderTest.java
@@ -1,4 +1,4 @@
-package com.alexstyl.specialdates.service;
+package com.alexstyl.specialdates.events.peopleevents;
 
 import com.alexstyl.specialdates.Optional;
 import com.alexstyl.specialdates.TestContactEventsBuilder;
@@ -10,12 +10,6 @@ import com.alexstyl.specialdates.date.Date;
 import com.alexstyl.specialdates.date.TimePeriod;
 import com.alexstyl.specialdates.events.namedays.NamedayUserSettings;
 import com.alexstyl.specialdates.events.namedays.calendar.resource.NamedayCalendarProvider;
-import com.alexstyl.specialdates.events.peopleevents.AndroidPeopleStaticEventsProvider;
-import com.alexstyl.specialdates.events.peopleevents.ContactEventsOnADate;
-import com.alexstyl.specialdates.events.peopleevents.CustomEventProvider;
-import com.alexstyl.specialdates.events.peopleevents.NoEventsFoundException;
-import com.alexstyl.specialdates.events.peopleevents.PeopleEventsProvider;
-import com.alexstyl.specialdates.events.peopleevents.PeopleNamedaysCalculator;
 import com.novoda.notils.logger.simple.Log;
 
 import java.util.Collections;
